+++ example/integration_test/plugin_integration_test.dart
@@ -15,8 +15,8 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('getPlatformVersion test', (WidgetTester tester) async {
-    final AndroidLaunchableApps plugin = AndroidLaunchableApps();
-    final String? version = await plugin.getPlatformVersion();
+    //final AndroidLaunchableApps plugin = AndroidLaunchableApps();
+    final String? version = await AndroidLaunchableApps.getPlatformVersion();
     // The version string depends on the host platform running the test, so
     // just assert that some non-empty string is returned.
     expect(version?.isNotEmpty, true);