+++ example/lib/android_launchable_apps_example.dart
@@ -1,11 +1,12 @@
-import 'package:android_launchable_apps/models.dart';
-import 'package:flutter/material.dart';
 import 'dart:async';
 
+import 'package:flutter/material.dart';
+
 import 'package:android_launchable_apps/android_launchable_apps.dart';
+import 'package:android_launchable_apps/models.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(const MaterialApp(home: MyApp()));
 }
 
 /// Example application to display android user launchable apps
@@ -18,8 +19,6 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  final _androidLaunchableAppsPlugin = AndroidLaunchableApps();
-
   bool isLoading = false;
   List<AndroidAppInfo> apps = [];
 
@@ -34,8 +33,7 @@ class _MyAppState extends State<MyApp> {
       isLoading = true;
     });
 
-    final appList =
-        await _androidLaunchableAppsPlugin.getLaunchableApplications();
+    final appList = await AndroidLaunchableApps.getLaunchableApplications();
     appList.sort((a, b) => a.displayName.compareTo(b.displayName));
 
     setState(() {
@@ -44,28 +42,96 @@ class _MyAppState extends State<MyApp> {
     });
   }
 
+  Future<void> _checkUsagePermission() async {
+    bool hasPermission = await AndroidLaunchableApps.isUsagePermissionGranted();
+    await _popMessage('Is Usage Permission Granted? $hasPermission');
+  }
+
+  Future<void> _requestUsagePermission() async {
+    await AndroidLaunchableApps.requestUsagePermission();
+  }
+
+  Future<void> _checkAppUsage(AndroidAppInfo app) async {
+    bool hasPermission = await AndroidLaunchableApps.isUsagePermissionGranted();
+    if (!hasPermission) {
+      await _popMessage(
+          'Usage Permission has not been granted. Click the "request" button in this example app.');
+      return;
+    }
+
+    final DateTime startDate = DateUtils.dateOnly(DateTime.now());
+    final DateTime endDate = DateTime(
+        startDate.year,
+        startDate.month,
+        startDate.day + 1,
+        0,
+        0,
+        0,
+        startDate.millisecond - startDate.millisecond - 1);
+    final usages = await AndroidLaunchableApps.queryEvents(startDate, endDate);
+    final packageUsages = usages.where((u) => u.packageName == app.packageName);
+    await _popMessage(
+        'Events of today for ${app.displayName} (${app.packageName})\n\n${packageUsages.map((e) => e.toString()).join('\n')}');
+  }
+
+  Future<void> _popMessage(String text) async {
+    await showDialog(
+      context: context,
+      builder: (dialogContext) {
+        return AlertDialog(
+          content: SingleChildScrollView(child: Text(text)),
+          actions: [
+            ElevatedButton(
+                onPressed: () => Navigator.pop(dialogContext),
+                child: const Text('OK')),
+          ],
+        );
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Launchable Apps Example'),
-          actions: [
-            IconButton(
-                tooltip: 'Refresh Apps',
-                onPressed: _refreshApps,
-                icon: const Icon(Icons.refresh_outlined)),
-          ],
-        ),
-        body: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Center(
-            child: (isLoading)
-                ? const CircularProgressIndicator()
-                : _buildAppsUI(),
-          ),
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Launchable Apps Example'),
+        actions: [
+          IconButton(
+              tooltip: 'Refresh Apps',
+              onPressed: _refreshApps,
+              icon: const Icon(Icons.refresh_outlined)),
+        ],
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: Center(
+          child:
+              (isLoading) ? const CircularProgressIndicator() : _buildContent(),
         ),
       ),
+    );
+  }
+
+  Widget _buildContent() {
+    return Column(
+      spacing: 10,
+      children: [
+        _buildUsageUI(),
+        Expanded(child: _buildAppsUI()),
+      ],
+    );
+  }
+
+  Widget _buildUsageUI() {
+    return Wrap(
+      children: [
+        ElevatedButton(
+            onPressed: _checkUsagePermission,
+            child: const Text('Check Usage Permission')),
+        ElevatedButton(
+            onPressed: _requestUsagePermission,
+            child: const Text('Request Usage Permission')),
+      ],
     );
   }
 
@@ -76,9 +142,9 @@ class _MyAppState extends State<MyApp> {
       itemBuilder: (context, index) {
         final app = apps[index];
         return Row(
+          spacing: 10,
           children: [
             Image.memory(app.iconBytes, height: 35, width: 35),
-            const SizedBox(width: 10),
             Expanded(
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
@@ -88,7 +154,10 @@ class _MyAppState extends State<MyApp> {
                   Text(app.categoryName),
                 ],
               ),
-            )
+            ),
+            ElevatedButton(
+                onPressed: () => _checkAppUsage(app),
+                child: const Text('Usage')),
           ],
         );
       },