+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-## 0.0.1
+## 0.1.0
 
-* Describe initial release.
+* **Breaking change (Minor)** Change methods of AndroidLaunchableApps to be static
+* Add the ability to query for package event stats
+* Add check and request permission to query for package event stats 
+* Update example application to use the above new features.
 
 ## 0.0.2
 
 * Add example application.
 * Improve pub score with documentation and formatting.
+
+## 0.0.1
+
+* Describe initial release.